refactoring consts
