+++ gendiff/formaters/__init__.py
@@ -1,6 +1,7 @@
-from gendiff.formaters.main import get_output_string
+from gendiff.formaters.main import get_output_string, STYLISH_FORMAT
 
 
 __all__ = (
     'get_output_string',
+    'STYLISH_FORMAT',
 )
